third_party/WebKit/LayoutTests/platform/mac-lion/fast/js/stack-overflow-arrity-catch-expected.txt

Test that if an arrity check causes a stack overflow, the exception goes to the right catch

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gotRightCatch is true
FAIL gotWrongCatch1 should be false. Was true.
FAIL gotWrongCatch2 should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
